fix check
