+++ src/cmd_all/src/lib.rs
@@ -14,8 +14,9 @@
 
 #![feature(lazy_cell)]
 
-pub mod playground;
-pub use playground::*;
 mod all_in_one;
+mod common;
+pub mod playground;
 
 pub use all_in_one::*;
+pub use playground::*;
