+++ src/cmd_all/src/playground.rs
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 use std::env;
-use std::ffi::OsString;
 use std::io::Write;
 use std::path::Path;
 use std::sync::LazyLock;
@@ -25,28 +24,7 @@ use tokio::io::{AsyncBufReadExt, BufReader};
 use tokio::process::Command;
 use tokio::signal;
 
-pub enum RisingWaveService {
-    Compute(Vec<OsString>),
-    Meta(Vec<OsString>),
-    Frontend(Vec<OsString>),
-    Compactor(Vec<OsString>),
-    ConnectorNode(Vec<OsString>),
-}
-
-impl RisingWaveService {
-    /// Extend additional arguments to the service.
-    fn extend_args(&mut self, args: &[&str]) {
-        match self {
-            RisingWaveService::Compute(args0)
-            | RisingWaveService::Meta(args0)
-            | RisingWaveService::Frontend(args0)
-            | RisingWaveService::Compactor(args0)
-            | RisingWaveService::ConnectorNode(args0) => {
-                args0.extend(args.iter().map(|s| s.into()))
-            }
-        }
-    }
-}
+use crate::common::{osstrs, RisingWaveService};
 
 const IDLE_EXIT_SECONDS: u64 = 1800;
 
@@ -168,10 +146,6 @@ fn get_services(profile: &str) -> (Vec<RisingWaveService>, bool) {
         })
     }
     (services, idle_exit)
-}
-
-fn osstrs<const N: usize>(s: [&str; N]) -> Vec<OsString> {
-    s.iter().map(OsString::from).collect()
 }
 
 #[derive(Debug, Clone, Parser)]
